Feat: Create a conversation before uploading files #3221
