+++ web/src/pages/next-chats/hooks/use-create-conversation.ts
@@ -1,13 +1,11 @@
-import { useGetChatSearchParams } from '@/hooks/use-chat-request';
 import { useCallback } from 'react';
-import {
-  useSetChatRouteParams,
-  useSetConversation,
-} from './use-send-chat-message';
+import { useParams } from 'umi';
+import { useSetChatRouteParams } from './use-set-chat-route';
+import { useSetConversation } from './use-set-conversation';
 
 export const useCreateConversationBeforeUploadDocument = () => {
   const { setConversation } = useSetConversation();
-  const { dialogId } = useGetChatSearchParams();
+  const { id: dialogId } = useParams();
   const { getConversationIsNew } = useSetChatRouteParams();
 
   const createConversationBeforeUploadDocument = useCallback(
